requirements.txt
langchain
wikipedia
langchain-community
langchain-google-genai
python-dotenv
pydantic
duckduckgo-search
Flask
gunicorn # Often useful for production WSGI server
